Emotion-and-Behavior-Analysis-of-Viewers-During-Advertisement-Watching
This study focuses on a project developed to measure the impact of commercials and analyze the emotional reactions of the audience. In the project, a model was developed to detect the emotional reactions of viewers to commercials. The model is a deep learning network created using a convolutional neural network. After watching the commercial, the emotional reactions of the audience are visualized with graphics. These graphics allow the classification of the emotions that viewers give to the commercial, the analysis of emotional transitions, and the evaluation of the features of the commercial. This method is used to objectively evaluate the impact of commercials and develop advertising strategies.